Add files via upload
This is a simple to-do-item application in Java. It supports the following features:
Add a to-do item
Delete a to-do item
View the to-do items
Quit and exit list
Description:
Make sure to structure your program in a modular way. In this case, that means you would have a command-line application which uses a class that holds the to-do items internally and provides public methods to add an item, delete an item, and provide the list of to-do items.